+++ changelog.md
@@ -4,10 +4,13 @@
   - Update some dev dependencies: Tailwindcss, GH Pages, TS ESLint
   - Refactor `FetchRequest` method parameters to options-based
   - Request sent to `DPNS` when/if push subscription revoked by end-user
+  - Fix endless logout loop if 401 is returned
 - UI:
   - Enable notification icon
   - Add random hint in main menu
-  - Remove Valentine's animation :'(
+  - Remove Valentine's animation
+- Authentication:
+  - Add Microsoft Entra ID option
 
 # 0.8.8.2 Beta Valentine's Edition - 10/02/2026 20:10 UTC+1